Move delete operations to delete.go
